Move shared bits of code
